+++ src/rpc/misc.cpp
@@ -92,9 +92,9 @@ UniValue getinfo(const UniValue& params, bool fHelp)
     obj.pushKV("subversion", strSubVersion);
     obj.pushKV("protocolversion", PROTOCOL_VERSION);
 #ifdef ENABLE_WALLET
-    if (pwalletMain) {
+    if (pwalletMain && fEnableGetinfoWalletMetadata) {
         obj.pushKV("walletversion", pwalletMain->GetVersion());
-        obj.pushKV("balance",       ValueFromAmount(pwalletMain->GetBalance(std::nullopt)));
+        obj.pushKV("balance",       ValueFromAmount(pwalletMain->GetBalance(std::nullopt, ISMINE_SPENDABLE_ANY, 0)));
     }
 #endif
     obj.pushKV("blocks",        (int)chainActive.Height());
@@ -104,13 +104,15 @@ UniValue getinfo(const UniValue& params, bool fHelp)
     obj.pushKV("difficulty",    (double)GetDifficulty());
     obj.pushKV("testnet",       Params().TestnetToBeDeprecatedFieldRPC());
 #ifdef ENABLE_WALLET
-    if (pwalletMain) {
-        obj.pushKV("keypoololdest", pwalletMain->GetOldestKeyPoolTime());
-        obj.pushKV("keypoolsize",   (int)pwalletMain->GetKeyPoolSize());
+    if (fEnableGetinfoWalletMetadata) {
+        if (pwalletMain) {
+            obj.pushKV("keypoololdest", pwalletMain->GetOldestKeyPoolTime());
+            obj.pushKV("keypoolsize",   (int)pwalletMain->GetKeyPoolSize());
+        }
+        if (pwalletMain && pwalletMain->IsCrypted())
+            obj.pushKV("unlocked_until", nWalletUnlockTime);
+        obj.pushKV("paytxfee",      ValueFromAmount(payTxFee.GetFeePerK()));
     }
-    if (pwalletMain && pwalletMain->IsCrypted())
-        obj.pushKV("unlocked_until", nWalletUnlockTime);
-    obj.pushKV("paytxfee",      ValueFromAmount(payTxFee.GetFeePerK()));
 #endif
     obj.pushKV("relayfee",      ValueFromAmount(::minRelayTxFee.GetFeePerK()));
     auto warnings = GetWarnings("statusbar");
@@ -207,7 +209,7 @@ UniValue validateaddress(const UniValue& params, bool fHelp)
 
 #ifdef ENABLE_WALLET
         isminetype mine = pwalletMain ? IsMine(*pwalletMain, dest) : ISMINE_NO;
-        ret.pushKV("ismine", (mine & ISMINE_SPENDABLE) ? true : false);
+        ret.pushKV("ismine", (mine & ISMINE_SPENDABLE_ANY) ? true : false);
         ret.pushKV("iswatchonly", (mine & ISMINE_WATCH_ONLY) ? true: false);
         UniValue detail = std::visit(DescribeAddressVisitor(), dest);
         ret.pushKVs(detail);