+++ src/wallet/rpcwallet.cpp
@@ -118,6 +118,13 @@ void ThrowIfInitialBlockDownload()
     }
 }
 
+UniValue BalancesObject(CAmount balance) {
+    UniValue obj(UniValue::VOBJ);
+    obj.pushKV("value", FormatMoney(balance));
+    obj.pushKV("valueZat", balance);
+    return obj;
+}
+
 void WalletTxToJSON(const CWalletTx& wtx, UniValue& entry, const std::optional<int>& asOfHeight)
 {
     int confirms = wtx.GetDepthInMainChain(asOfHeight);
@@ -267,7 +274,7 @@ UniValue getrawchangeaddress(const UniValue& params, bool fHelp)
 
 static void SendMoney(const CTxDestination &address, CAmount nValue, bool fSubtractFeeFromAmount, CWalletTx& wtxNew)
 {
-    CAmount curBalance = pwalletMain->GetBalance(std::nullopt);
+    CAmount curBalance = pwalletMain->GetBalance(std::nullopt, ISMINE_LEGACY_SPENDABLE, 0);
 
     // Check amount
     if (nValue <= 0)
@@ -288,7 +295,7 @@ static void SendMoney(const CTxDestination &address, CAmount nValue, bool fSubtr
     CRecipient recipient = {scriptPubKey, nValue, fSubtractFeeFromAmount};
     vecSend.push_back(recipient);
     if (!pwalletMain->CreateTransaction(vecSend, wtxNew, reservekey, nFeeRequired, nChangePosRet, strError)) {
-        if (!fSubtractFeeFromAmount && nValue + nFeeRequired > pwalletMain->GetBalance(std::nullopt))
+        if (!fSubtractFeeFromAmount && nValue + nFeeRequired > pwalletMain->GetBalance(std::nullopt, ISMINE_LEGACY_SPENDABLE, 0))
             strError = strprintf("Error: This transaction requires a transaction fee of at least %s because of its amount, complexity, or use of recently received funds!", FormatMoney(nFeeRequired));
         throw JSONRPCError(RPC_WALLET_ERROR, strError);
     }
@@ -1102,9 +1109,12 @@ UniValue getbalance(const UniValue& params, bool fHelp)
     if (fHelp || params.size() > 5)
         throw runtime_error(
             "getbalance ( \"(dummy)\" minconf includeWatchonly inZat asOfHeight )\n"
-            "\nReturns the wallet's available transparent balance. This total\n"
-            "currently includes transparent balances associated with unified\n"
-            "accounts. Prefer to use `z_getbalanceforaccount` instead.\n"
+            "\nReturns the wallet's available balance associated with legacy\n"
+            "transparent addresses. Prefer to use `z_getbalanceforaccount` when\n"
+            "possible.\n"
+            "NOTE: starting in zcashd v4.7.0, this method incorrectly returned\n"
+            "transparent value associated with transparent receivers of the wallet's\n"
+            "unified addresses. This error was corrected in zcashd v5.6.0.\n"
             "\nArguments:\n"
             "1. (dummy)          (string, optional) Remains for backward compatibility. Must be excluded or set to \"*\" or \"\".\n"
             "2. minconf          (numeric, optional, default=0) Only include transactions confirmed at least this many times.\n"
@@ -1135,7 +1145,7 @@ UniValue getbalance(const UniValue& params, bool fHelp)
 
     int min_depth = parseMinconf(0, params, 1, asOfHeight);
 
-    isminefilter filter = ISMINE_SPENDABLE;
+    isminefilter filter = ISMINE_LEGACY_SPENDABLE;
     if (!params[2].isNull() && params[2].get_bool()) {
         filter = filter | ISMINE_WATCH_ONLY;
     }
@@ -1160,7 +1170,7 @@ UniValue getunconfirmedbalance(const UniValue &params, bool fHelp)
 
     LOCK2(cs_main, pwalletMain->cs_wallet);
 
-    return ValueFromAmount(pwalletMain->GetUnconfirmedTransparentBalance());
+    return ValueFromAmount(pwalletMain->GetUnconfirmedTransparentBalance(ISMINE_LEGACY_SPENDABLE));
 }
 
 
@@ -1263,7 +1273,7 @@ UniValue sendmany(const UniValue& params, bool fHelp)
     EnsureWalletIsUnlocked();
 
     // Check funds
-    if (totalAmount > pwalletMain->GetLegacyBalance(ISMINE_SPENDABLE, nMinDepth)) {
+    if (totalAmount > pwalletMain->GetLegacyBalance(ISMINE_LEGACY_SPENDABLE, nMinDepth)) {
         throw JSONRPCError(RPC_WALLET_INSUFFICIENT_FUNDS, "Wallet has insufficient funds");
     }
 
@@ -1372,7 +1382,7 @@ UniValue ListReceived(const UniValue& params)
     if (params.size() > 1)
         fIncludeEmpty = params[1].get_bool();
 
-    isminefilter filter = ISMINE_SPENDABLE;
+    isminefilter filter = ISMINE_LEGACY_SPENDABLE;
     if(params.size() > 2)
         if(params[2].get_bool())
             filter = filter | ISMINE_WATCH_ONLY;
@@ -1459,9 +1469,9 @@ UniValue listreceivedbyaddress(const UniValue& params, bool fHelp)
         throw runtime_error(
             "listreceivedbyaddress ( minconf includeempty includeWatchonly addressFilter includeImmatureCoinbase asOfHeight )\n"
             "\nList balances by transparent receiving address. This API does not provide\n"
-            "any information for associated with shielded addresses and should only be used\n"
-            "in circumstances where it is necessary to interoperate with legacy Bitcoin\n"
-            "infrastructure.\n"
+            "any information for funds associated with shielded addresses or transparent\n"
+            "receivers of unified addresses and should only be used in circumstances where\n"
+            "it is necessary to interoperate with legacy Bitcoin infrastructure.\n"
             "\nArguments:\n"
             "1. minconf       (numeric, optional, default=1) The minimum number of confirmations before payments are included.\n"
             "2. includeempty  (numeric, optional, default=false) Whether to include addresses that haven't received any payments.\n"
@@ -1592,9 +1602,10 @@ UniValue listtransactions(const UniValue& params, bool fHelp)
             "\nReturns up to 'count' of the most recent transactions associated with legacy transparent\n"
             "addresses of this wallet, skipping the first 'from' transactions.\n"
             "\nThis API does not provide any information about transactions containing shielded inputs\n"
-            "or outputs, and should only be used in circumstances where it is necessary to interoperate\n"
-            "with legacy Bitcoin infrastructure. Use z_listreceivedbyaddress to obtain information about\n"
-            "the wallet's shielded transactions.\n"
+            "or outputs or associated with transparent receivers of unified addresses, and should only\n"
+            "be used in circumstances where it is necessary to interoperate with legacy Bitcoin\n"
+            "infrastructure. Use z_listreceivedbyaddress to obtain information about the wallet's \n"
+            "shielded transactions.\n"
             "\nArguments:\n"
             "1. (dummy)        (string, optional) If set, should be \"*\" for backwards compatibility.\n"
             "2. count          (numeric, optional, default=10) The number of transactions to return\n"
@@ -1650,7 +1661,7 @@ UniValue listtransactions(const UniValue& params, bool fHelp)
     int nFrom = 0;
     if (params.size() > 2)
         nFrom = params[2].get_int();
-    isminefilter filter = ISMINE_SPENDABLE;
+    isminefilter filter = ISMINE_LEGACY_SPENDABLE;
     if(params.size() > 3)
         if(params[3].get_bool())
             filter = filter | ISMINE_WATCH_ONLY;
@@ -1713,6 +1724,10 @@ UniValue listsinceblock(const UniValue& params, bool fHelp)
         throw runtime_error(
             "listsinceblock ( \"blockhash\" target-confirmations includeWatchonly includeRemoved includeChange asOfHeight )\n"
             "\nGet all transactions in blocks since block [blockhash], or all transactions if omitted\n"
+            "\nThis API does not provide any information about transactions containing shielded inputs\n"
+            "or outputs or associated with transparent receivers of unified addresses, and should only\n"
+            "be used in circumstances where it is necessary to interoperate with legacy Bitcoin\n"
+            "infrastructure.\n"
             "\nArguments:\n"
             "1. \"blockhash\"   (string, optional) The block hash to list transactions since\n"
             "2. target-confirmations:    (numeric, optional) The confirmations required, must be 1 or more\n"
@@ -1758,7 +1773,7 @@ UniValue listsinceblock(const UniValue& params, bool fHelp)
 
     CBlockIndex *pindex = NULL;
     int target_confirms = 1;
-    isminefilter filter = ISMINE_SPENDABLE;
+    isminefilter filter = ISMINE_LEGACY_SPENDABLE;
 
     if (params.size() > 0)
     {
@@ -1827,8 +1842,10 @@ UniValue gettransaction(const UniValue& params, bool fHelp)
         throw runtime_error(
             "gettransaction \"txid\" ( includeWatchonly verbose asOfHeight )\n"
             "\nReturns detailed information about in-wallet transaction <txid>. This does not\n"
-            "include complete information about shielded components of the transaction; to obtain\n"
-            "details about shielded components of the transaction use `z_viewtransaction`.\n"
+            "include complete information about shielded components of the transaction, and it\n"
+            "does not include any value associated with transparent receivers of unified addresses;\n"
+            "to obtain details about a transaction with shielded component, or received at a\n"
+            "unified address use `z_viewtransaction`.\n"
             "\nArguments:\n"
             "1. \"txid\"    (string, required) The transaction id\n"
             "2. includeWatchonly    (bool, optional, default=false) Whether to include watchonly addresses in balance calculation and details[]\n"
@@ -1850,7 +1867,7 @@ UniValue gettransaction(const UniValue& params, bool fHelp)
             "  \"details\" : [\n"
             "    {\n"
             "      \"address\" : \"zcashaddress\",   (string) The Zcash address involved in the transaction\n"
-            "      \"category\" : \"send|receive\",    (string) The category, either 'send' or 'receive'\n"
+            "      \"category\" : \"send|receive\",  (string) The category, either 'send' or 'receive'\n"
             "      \"amount\" : x.xxx                  (numeric) The amount in " + CURRENCY_UNIT + "\n"
             "      \"amountZat\" : x                   (numeric) The amount in " + MINOR_CURRENCY_UNIT + "\n"
             "      \"vout\" : n,                       (numeric) the vout value\n"
@@ -1883,7 +1900,7 @@ UniValue gettransaction(const UniValue& params, bool fHelp)
     uint256 hash;
     hash.SetHex(params[0].get_str());
 
-    isminefilter filter = ISMINE_SPENDABLE;
+    isminefilter filter = ISMINE_LEGACY_SPENDABLE;
     if(params.size() > 1)
         if(params[1].get_bool())
             filter = filter | ISMINE_WATCH_ONLY;
@@ -2444,13 +2461,17 @@ UniValue getwalletinfo(const UniValue& params, bool fHelp)
             "\nResult:\n"
             "{\n"
             "  \"walletversion\": xxxxx,     (numeric) the wallet version\n"
-            "  \"balance\": xxxxxxx,         (numeric) the total confirmed transparent balance of the wallet in " + CURRENCY_UNIT + "\n"
-            "  \"unconfirmed_balance\": xxx, (numeric, optional) the total unconfirmed transparent balance of the wallet in " + CURRENCY_UNIT + ".\n"
-            "                              Not included if `asOfHeight` is specified.\n"
-            "  \"immature_balance\": xxxxxx, (numeric) the total immature transparent balance of the wallet in " + CURRENCY_UNIT + "\n"
-            "  \"shielded_balance\": xxxxxxx,  (numeric) the total confirmed shielded balance of the wallet in " + CURRENCY_UNIT + "\n"
-            "  \"shielded_unconfirmed_balance\": xxx, (numeric, optional) the total unconfirmed shielded balance of the wallet in " + CURRENCY_UNIT + ".\n"
-            "                              Not included if `asOfHeight` is specified.\n"
+            "  \"balance\": xxxxxxx,         (numeric, deprecated) the total confirmed transparent balance of the wallet in " + CURRENCY_UNIT + ".\n"
+            "                              This includes all value controlled by keys in the wallet, irrespective of what unified \n"
+            "                              account the funds may be associated with (if any). Deprecated, use `z_getbalances` instead.\n"
+            "  \"unconfirmed_balance\": xxx, (numeric, optional, deprecated) the total unconfirmed transparent balance of the wallet in " + CURRENCY_UNIT + ".\n"
+            "                              Not included if `asOfHeight` is specified. Depreceated, use `z_getbalances` instead.\n"
+            "  \"immature_balance\": xxxxxxx,  (numeric, deprecated) the total immature transparent balance of the wallet in " + CURRENCY_UNIT + ".\n"
+            "                              Deprecated, use `z_getbalances` instead.\n"
+            "  \"shielded_balance\": xxxxxxx,  (numeric, deprecated) the total confirmed shielded balance of the wallet in " + CURRENCY_UNIT + ".\n"
+            "                              Deprecated, use `z_getbalances` instead.\n"
+            "  \"shielded_unconfirmed_balance\": xxx, (numeric, optional, deprecated) the total unconfirmed shielded balance of the wallet in " + CURRENCY_UNIT + ".\n"
+            "                              Not included if `asOfHeight` is specified. Deprecated, use `z_getbalances` instead.\n"
             "  \"txcount\": xxxxxxx,         (numeric) the total number of transactions in the wallet\n"
             "  \"keypoololdest\": xxxxxx,    (numeric) the timestamp (seconds since GMT epoch) of the oldest pre-generated key in the key pool\n"
             "  \"keypoolsize\": xxxx,        (numeric) how many new keys are pre-generated\n"
@@ -2472,14 +2493,16 @@ UniValue getwalletinfo(const UniValue& params, bool fHelp)
 
     UniValue obj(UniValue::VOBJ);
     obj.pushKV("walletversion", pwalletMain->GetVersion());
-    obj.pushKV("balance",       ValueFromAmount(pwalletMain->GetBalance(asOfHeight)));
-    if (!asOfHeight.has_value()) {
-        obj.pushKV("unconfirmed_balance", ValueFromAmount(pwalletMain->GetUnconfirmedTransparentBalance()));
-    }
-    obj.pushKV("immature_balance",    ValueFromAmount(pwalletMain->GetImmatureBalance(asOfHeight)));
-    obj.pushKV("shielded_balance",    FormatMoney(getBalanceZaddr(std::nullopt, asOfHeight, 1, INT_MAX)));
-    if (!asOfHeight.has_value()) {
-        obj.pushKV("shielded_unconfirmed_balance", FormatMoney(getBalanceZaddr(std::nullopt, asOfHeight, 0, 0)));
+    if (fEnableWalletinfoBalances) {
+        obj.pushKV("balance",       ValueFromAmount(pwalletMain->GetBalance(asOfHeight, ISMINE_SPENDABLE_ANY, 0)));
+        if (!asOfHeight.has_value()) {
+            obj.pushKV("unconfirmed_balance", ValueFromAmount(pwalletMain->GetUnconfirmedTransparentBalance(ISMINE_SPENDABLE_ANY)));
+        }
+        obj.pushKV("immature_balance",    ValueFromAmount(pwalletMain->GetImmatureBalance(asOfHeight)));
+        obj.pushKV("shielded_balance",    FormatMoney(getBalanceZaddr(std::nullopt, asOfHeight, 1, INT_MAX)));
+        if (!asOfHeight.has_value()) {
+            obj.pushKV("shielded_unconfirmed_balance", FormatMoney(getBalanceZaddr(std::nullopt, asOfHeight, 0, 0)));
+        }
     }
     obj.pushKV("txcount",       (int)pwalletMain->mapWallet.size());
     obj.pushKV("keypoololdest", pwalletMain->GetOldestKeyPoolTime());
@@ -2608,15 +2631,13 @@ UniValue listunspent(const UniValue& params, bool fHelp)
     LOCK2(cs_main, pwalletMain->cs_wallet);
 
     UniValue results(UniValue::VARR);
-    vector<COutput> vecOutputs;
-    pwalletMain->AvailableCoins(
-            vecOutputs,
+    vector<COutput> vecOutputs = pwalletMain->AvailableCoins(
             asOfHeight,
             false,        // fOnlyConfirmed
             nullptr,      // coinControl
             true,         // fIncludeZeroValue
             true,         // fIncludeCoinBase
-            false,        // fOnlySpendable
+            ISMINE_LEGACY_SPENDABLE,
             nMinDepth,
             destinations);
     for (const COutput& out : vecOutputs) {
@@ -3557,17 +3578,20 @@ UniValue z_listunifiedreceivers(const UniValue& params, bool fHelp)
 }
 
 CAmount getBalanceTaddr(const std::optional<CTxDestination>& taddr, const std::optional<int>& asOfHeight, int minDepth=1, bool ignoreUnspendable=true) {
-    vector<COutput> vecOutputs;
     CAmount balance = 0;
 
     LOCK2(cs_main, pwalletMain->cs_wallet);
 
-    pwalletMain->AvailableCoins(vecOutputs, asOfHeight, false, NULL, true);
-    for (const COutput& out : vecOutputs) {
-        if (out.nDepth < minDepth) {
-            continue;
-        }
+    vector<COutput> vecOutputs = pwalletMain->AvailableCoins(
+            asOfHeight,
+            false,        // fOnlyConfirmed
+            nullptr,      // coinControl
+            true,         // fIncludeZeroValue
+            true,         // fIncludeCoinBase
+            ignoreUnspendable ? ISMINE_LEGACY_SPENDABLE : ISMINE_LEGACY_ALL,
+            minDepth);
 
+    for (const COutput& out : vecOutputs) {
         if (ignoreUnspendable && !out.fSpendable) {
             continue;
         }
@@ -4229,6 +4253,244 @@ UniValue z_gettotalbalance(const UniValue& params, bool fHelp)
     result.pushKV("transparent", FormatMoney(nBalance));
     result.pushKV("private", FormatMoney(nPrivateBalance));
     result.pushKV("total", FormatMoney(nTotalBalance));
+    return result;
+}
+
+UniValue z_getbalances(const UniValue& params, bool fHelp)
+{
+    if (!EnsureWalletIsAvailable(fHelp))
+        return NullUniValue;
+
+    if (fHelp || params.size() > 2)
+        throw runtime_error(
+            "z_getbalances ( minconf )\n"
+            "\nReturns the balances available for each independent spending authority held by the wallet,\n"
+            "and optionally the balances and received amounts associated with imported watch-only addresses\n"
+            "and viewing keys.\n"
+            "This includes the legacy transparent pool of funds, each distinct Sprout and Sapling address\n"
+            "derived using `z_getnewaddress` or from spending keys imported with `z_importkey`, and funds\n"
+            "held by each HD-derived Unified Account in the wallet.\n"
+            "\nArguments:\n"
+            "1. minconf          (numeric, optional, default=1) Only include private and transparent transactions confirmed at least this many times.\n"
+            "2. includeWatchonly (bool, optional, default=false) Also include balance in watchonly addresses (see 'importaddress' and 'z_importviewingkey')\n"
+            "\nResult:\n"
+            "{\n"
+            "  \"legacy_transparent\": {                (object, optional) the balances of transparent funds held by legacy transparent keys,\n"
+            "                                         in " + CURRENCY_UNIT + " and " + MINOR_CURRENCY_UNIT + ". The object format documented\n"
+            "                                         for this value is also used for each object denoted by `{...}` below.\n"
+            "                                         All funds held in legacy transparent addresses are treated as though they are associated\n"
+            "                                         with a single spending authority.\n"
+            "    \"value\": xxxxx,\n                    (numeric) the amount in " + CURRENCY_UNIT + "\n"
+            "    \"valueZat\": xxxxx,\n                 (numeric) the amount in " + MINOR_CURRENCY_UNIT + "\n"
+            "  },\n"
+            "  \"transparent_watchonly\": {             (object, optional) the balances of transparent funds associated with watch-only transparent addresses\n"
+            "                                         Each such address is treated as being associated with an independent spending authority.\n"
+            "    \"<t_addr_1>\": {...},\n"
+            "    \"<t_addr_2>\": {...},\n"
+            "    ...\n"
+            "  },\n"
+            "  \"sprout\": {                            (object, optional) the balances held by each Sprout spending key in the wallet, keyed by address.\n"
+            "                                         Each such address is treated as being associated with an independent spending authority.\n"
+            "    \"<sprout_addr_1>\": {...},\n"
+            "    \"<sprout_addr_2>\": {...},\n"
+            "    ...\n"
+            "  },\n"
+            "  \"sprout_watchonly_received\": {         (object, optional) the received total visible to each Sprout viewing key in the wallet where the wallet\n"
+            "                                         has no corresponding spending authority, keyed by address. Each such address is treated as being \n"
+            "                                         associated with an independent spending authority. Note that these do not represent balances, as Sprout\n"
+            "                                         viewing keys are not capable of detecting when notes are spent.\n"
+            "    \"<sprout_addr_1>\": {...}\n"
+            "    ...\n"
+            "  },\n"
+            "  \"legacy_sapling\": {                    (object, optional) the balances held by each legacy Sapling spending key in the wallet, keyed by address.\n"
+            "                                           Each such address is treated as being associated with an independent spending authority.\n"
+            "    \"<sapling_addr_1>\": {...},\n"
+            "    \"<sapling_addr_2>\": {...},\n"
+            "    ...\n"
+            "  },\n"
+            "  \"sapling_watchonly\": {                 (object, optional) the balance viewable by each Sapling viewing key in the wallet for which the wallet\n"
+            "                                           does not hold spending authority, keyed by the default address corresponding to the associated incoming\n"
+            "                                           viewing key. Each such address is associated with an independent spending authority. \n"
+            "    \"<sapling_addr_1>\": {...},\n"
+            "    \"<sapling_addr_2>\": {...},\n"
+            "    ...\n"
+            "  },\n"
+            "  \"accounts\": [                          (object, optional) the balances held by each Unified Account spending authority in the wallet\n"
+            "    {\n"
+            "      \"account\": xxx, n                  (numeric) the Unified Account identifier\n"
+            "      \"transparent\": {...},              (numeric, optional) the transparent balance held by the account\n"
+            "      \"sapling\": {...},                  (numeric, optional) the Sapling balance held by the account\n"
+            "      \"orchard\": {...},                  (numeric, optional) the Orchard balance held by the account\n"
+            "      \"total\": {...}                     (numeric) the total funds in all pools held by the account\n"
+            "    }\n"
+            "    ...\n"
+            "  ],\n"
+            "  \"total\": {...},                        (object) the total of all funds for which this wallet controls spending keys,\n"
+            "}\n"
+            "\nExamples:\n"
+            "\nThe totals of all funds held by (and optionally watched by) keys in the wallet\n"
+            + HelpExampleCli("z_getbalances", "") +
+            "\nThe totals of all funds held by the wallet at least 5 blocks confirmed\n"
+            + HelpExampleCli("z_getbalances", "5") +
+            "\nAs a JSON RPC call\n"
+            + HelpExampleRpc("z_getbalances", "5")
+        );
+
+    LOCK2(cs_main, pwalletMain->cs_wallet);
+
+    int nMinDepth = parseMinconf(1, params, 0, std::nullopt);
+
+    KeyIO keyIO(Params());
+    CAmount nTotalBalance = 0;
+
+    // legacy transparent balance
+    bool hasLegacyTransparent = false;
+    CAmount nLegacyTransparentBalance = 0;
+    vector<COutput> legacyTransparentOutputs = pwalletMain->AvailableCoins(
+            std::nullopt,
+            false,       // fOnlyConfirmed
+            nullptr,     // coinControl
+            false,       // fIncludeZeroValue
+            true,        // fIncludeCoinBase
+            ISMINE_LEGACY_SPENDABLE,
+            nMinDepth);
+    for (const COutput& out : legacyTransparentOutputs) {
+        hasLegacyTransparent = true;
+        nLegacyTransparentBalance += out.tx->vout[out.i].nValue;
+        nTotalBalance += out.tx->vout[out.i].nValue;
+    }
+
+    // transparent watch-only balances
+    bool hasTransparentWatchOnly = false;
+    UniValue transparentWatchOnlyBalances(UniValue::VOBJ);
+    vector<COutput> transparentWatchOnlyOutputs = pwalletMain->AvailableCoins(
+            std::nullopt,
+            false,       // fOnlyConfirmed
+            nullptr,     // coinControl
+            false,       // fIncludeZeroValue
+            true,        // fIncludeCoinBase
+            ISMINE_WATCH_ONLY,
+            nMinDepth);
+    for (const COutput& out : transparentWatchOnlyOutputs) {
+        hasTransparentWatchOnly = true;
+        CTxDestination addr;
+        if (ExtractDestination(out.TxOut().scriptPubKey, addr)) {
+            transparentWatchOnlyBalances.pushKV(
+                    keyIO.EncodeDestination(addr),
+                    BalancesObject(out.Value()));
+        } else {
+            // It's not clear that we could ever encounter this condition given,
+            // how address imports work, but we handle it for completeness.
+            transparentWatchOnlyBalances.pushKV(
+                    "(Address Unknown)",
+                    BalancesObject(out.Value()));
+        }
+    }
+
+    /// sprout balances
+    bool hasSprout = false;
+    bool hasSproutWatchOnly = false;
+    UniValue sproutBalances(UniValue::VOBJ);
+    UniValue sproutWatchOnlyReceived(UniValue::VOBJ);
+    std::set<libzcash::SproutPaymentAddress> sproutAddresses;
+    pwalletMain->GetSproutPaymentAddresses(sproutAddresses);
+    for (const SproutPaymentAddress& addr : sproutAddresses) {
+        auto sproutBalance = getBalanceZaddr(addr, std::nullopt, nMinDepth, INT_MAX, false);
+        if (pwalletMain->HaveSproutSpendingKey(addr)) {
+            hasSprout = true;
+            sproutBalances.pushKV(keyIO.EncodePaymentAddress(addr), BalancesObject(sproutBalance));
+            nTotalBalance += sproutBalance;
+        } else {
+            hasSproutWatchOnly = true;
+            sproutWatchOnlyReceived.pushKV(keyIO.EncodePaymentAddress(addr), BalancesObject(sproutBalance));
+        }
+    }
+
+    /// legacy sapling & sapling watch-only addresses
+    bool hasLegacySapling = false;
+    bool hasSaplingWatchOnly = false;
+    UniValue legacySaplingBalances(UniValue::VOBJ);
+    UniValue saplingWatchOnlyBalances(UniValue::VOBJ);
+    std::set<libzcash::SaplingPaymentAddress> legacySaplingAddrs;
+    pwalletMain->GetSaplingPaymentAddresses(legacySaplingAddrs);
+    for (const SaplingPaymentAddress& addr : legacySaplingAddrs) {
+        // if the address is associated with the legacy HD seed, or if there is no corresponding
+        // unified address metadata, then this is a legacy or watchonly address
+        if (GetSourceForPaymentAddress(pwalletMain)(addr) == PaymentAddressSource::LegacyHDSeed
+                    || !pwalletMain->GetUFVKMetadataForReceiver(addr).has_value()) {
+            auto saplingBalance = getBalanceZaddr(addr, std::nullopt, nMinDepth, INT_MAX, false);
+            if (pwalletMain->HaveSaplingSpendingKeyForAddress(addr)) {
+                hasLegacySapling = true;
+                legacySaplingBalances.pushKV(keyIO.EncodePaymentAddress(addr), BalancesObject(saplingBalance));
+                nTotalBalance += saplingBalance;
+            } else {
+                hasSaplingWatchOnly = true;
+                saplingWatchOnlyBalances.pushKV(keyIO.EncodePaymentAddress(addr), BalancesObject(saplingBalance));
+            }
+        }
+    }
+
+    // unified account balances
+    bool hasUnifiedAccounts = false;
+    UniValue accountBalances(UniValue::VARR);
+    for (const auto& [acctKey, ufvkId] : pwalletMain->mapUnifiedAccountKeys) {
+        hasUnifiedAccounts = true;
+        auto selector = pwalletMain->ZTXOSelectorForAccount(
+                acctKey.second,
+                true,
+                TransparentCoinbasePolicy::Allow
+                );
+        auto spendableInputs = pwalletMain->FindSpendableInputs(selector.value(), nMinDepth, std::nullopt);
+
+        CAmount transparentBalance = 0;
+        CAmount saplingBalance = 0;
+        CAmount orchardBalance = 0;
+        for (const auto& t : spendableInputs.utxos) {
+            transparentBalance += t.Value();
+        }
+        for (const auto& t : spendableInputs.saplingNoteEntries) {
+            saplingBalance += t.note.value();
+        }
+        for (const auto& t : spendableInputs.orchardNoteMetadata) {
+            orchardBalance += t.GetNoteValue();
+        }
+
+        CAmount accountTotal = transparentBalance + saplingBalance + orchardBalance;
+        nTotalBalance += accountTotal;
+
+        UniValue result(UniValue::VOBJ);
+        auto renderBalance = [&](std::string poolName, CAmount balance) {
+            if (balance > 0) {
+                result.pushKV(poolName, BalancesObject(balance));
+            }
+        };
+
+        const libzcash::AccountId accountId = acctKey.second;
+        result.pushKV("account", accountId);
+        renderBalance("transparent", transparentBalance);
+        renderBalance("sapling", saplingBalance);
+        renderBalance("orchard", orchardBalance);
+        renderBalance("total", accountTotal);
+
+        accountBalances.push_back(result);
+    }
+
+    UniValue result(UniValue::VOBJ);
+    if (hasLegacyTransparent)
+        result.pushKV("legacy_transparent", BalancesObject(nLegacyTransparentBalance));
+    if (hasTransparentWatchOnly)
+        result.pushKV("transparent_watchonly", transparentWatchOnlyBalances);
+    if (hasSprout)
+        result.pushKV("sprout", sproutBalances);
+    if (hasSproutWatchOnly)
+        result.pushKV("sprout_watchonly_received", sproutWatchOnlyReceived);
+    if (hasLegacySapling)
+        result.pushKV("legacy_sapling", legacySaplingBalances);
+    if (hasSaplingWatchOnly)
+        result.pushKV("sapling_watchonly", saplingWatchOnlyBalances);
+    if (hasUnifiedAccounts)
+        result.pushKV("accounts", accountBalances);
+    result.pushKV("total", BalancesObject(nTotalBalance));
     return result;
 }
 
@@ -5584,8 +5846,13 @@ UniValue z_mergetoaddress(const UniValue& params, bool fHelp)
 
     if (useAnyUTXO || taddrs.size() > 0) {
         // Get available utxos
-        vector<COutput> vecOutputs;
-        pwalletMain->AvailableCoins(vecOutputs, std::nullopt, true, NULL, false, false);
+        vector<COutput> vecOutputs = pwalletMain->AvailableCoins(
+            std::nullopt,
+            true,        // fOnlyConfirmed
+            nullptr,     // coinControl
+            false,       // fIncludeZeroValue
+            false,        // fIncludeCoinBase
+            ISMINE_SPENDABLE_ANY);
 
         // Find unspent utxos and update estimated size
         for (const COutput& out : vecOutputs) {
@@ -5967,6 +6234,7 @@ static const CRPCCommand commands[] =
     { "wallet",             "z_listreceivedbyaddress",  &z_listreceivedbyaddress,  false },
     { "wallet",             "z_listunspent",            &z_listunspent,            false },
     { "wallet",             "z_getbalance",             &z_getbalance,             false },
+    { "wallet",             "z_getbalances",            &z_getbalances,            false },
     { "wallet",             "z_gettotalbalance",        &z_gettotalbalance,        false },
     { "wallet",             "z_getbalanceforviewingkey",&z_getbalanceforviewingkey,false },
     { "wallet",             "z_getbalanceforaccount",   &z_getbalanceforaccount,   false },